Rename rwkv package to rwkv_cpp
