+++ python/chat_with_bot.py
@@ -10,7 +10,7 @@ import json
 import time
 import torch
 import sampling
-from rwkv import rwkv_cpp_shared_library, rwkv_cpp_model
+from rwkv_cpp import rwkv_cpp_shared_library, rwkv_cpp_model
 from tokenizer_util import get_tokenizer
 from typing import List, Dict, Optional
