+++ python/generate_completions.py
@@ -4,7 +4,7 @@
 import argparse
 import time
 import sampling
-from rwkv import rwkv_cpp_shared_library, rwkv_cpp_model
+from rwkv_cpp import rwkv_cpp_shared_library, rwkv_cpp_model
 from tokenizer_util import get_tokenizer
 
 # ======================================== Script settings ========================================
