+++ python/inference_example.py
@@ -3,7 +3,7 @@
 
 import argparse
 import sampling
-from rwkv import rwkv_cpp_shared_library, rwkv_cpp_model
+from rwkv_cpp import rwkv_cpp_shared_library, rwkv_cpp_model
 from tokenizer_util import get_tokenizer
 
 # Parse received arguments.
