+++ python/rwkv_cpp/rwkv_cpp_model.py
